LayoutTests/sputnik/Conformance/07_Lexical_Conventions/7.3_Line_Terminators/S7.3_A6_T3-expected.txt

CONSOLE MESSAGE: line 76: SyntaxError: Parse error
S7.3_A6_T3

PASS Expected parsing failure

TEST COMPLETE
